@@ -5,7 +5,7 @@
   "scripts": {
     "affected": "nx affected",
     "ng": "nx",
-    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2015 browser module main",
+    "postinstall": "node ./decorate-angular-cli.js && ngcc --properties es2020 browser module main",
     "nx": "nx",
     "start": "ng serve --hmr",
     "build": "ng build --configuration production",
@@ -31,19 +31,19 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.12",
-    "@angular/common": "13.3.12",
-    "@angular/compiler": "13.3.12",
-    "@angular/core": "13.3.12",
-    "@angular/forms": "13.3.12",
-    "@angular/platform-browser": "13.3.12",
-    "@angular/platform-browser-dynamic": "13.3.12",
-    "@angular/router": "13.3.12",
+    "@angular/animations": "14.0.7",
+    "@angular/common": "14.0.7",
+    "@angular/compiler": "14.0.7",
+    "@angular/core": "14.0.7",
+    "@angular/forms": "14.0.7",
+    "@angular/platform-browser": "14.0.7",
+    "@angular/platform-browser-dynamic": "14.0.7",
+    "@angular/router": "14.0.7",
     "@ngneat/svg-icon": "2.3.0",
-    "@ngrx/component-store": "13.0.2",
+    "@ngrx/component-store": "14.0.2",
     "@ngx-formly/core": "5.10.14",
-    "@nrwl/angular": "14.1.9",
-    "@nrwl/node": "14.1.9",
+    "@nrwl/angular": "14.4.3",
+    "@nrwl/node": "14.4.3",
     "immer": "8.0.1",
     "lodash.clonedeep": "4.5.0",
     "ng-indexed-db": "0.0.0-beta.1",
@@ -54,27 +54,27 @@
   },
   "devDependencies": {
     "@angular-builders/custom-webpack": "11.0.0",
-    "@angular-devkit/architect": "0.1303.10",
-    "@angular-devkit/build-angular": "13.3.10",
-    "@angular-eslint/eslint-plugin": "13.0.1",
-    "@angular-eslint/eslint-plugin-template": "13.0.1",
-    "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "13.3.10",
-    "@angular/compiler-cli": "13.3.12",
-    "@angular/language-service": "13.3.12",
+    "@angular-devkit/architect": "0.1400.7",
+    "@angular-devkit/build-angular": "14.0.7",
+    "@angular-eslint/eslint-plugin": "14.0.4",
+    "@angular-eslint/eslint-plugin-template": "14.0.4",
+    "@angular-eslint/template-parser": "14.0.4",
+    "@angular/cli": "~14.0.0",
+    "@angular/compiler-cli": "14.0.7",
+    "@angular/language-service": "14.0.7",
     "@hirez_io/observer-spy": "2.1.0",
     "@ngneat/spectator": "7.0.0",
     "@ngneat/tailwind": "6.0.3",
-    "@nrwl/cli": "14.1.9",
-    "@nrwl/cypress": "14.1.9",
-    "@nrwl/eslint-plugin-nx": "14.1.9",
-    "@nrwl/jest": "14.1.9",
-    "@nrwl/workspace": "14.1.9",
+    "@nrwl/cli": "14.4.3",
+    "@nrwl/cypress": "14.4.3",
+    "@nrwl/eslint-plugin-nx": "14.4.3",
+    "@nrwl/jest": "14.4.3",
+    "@nrwl/workspace": "14.4.3",
     "@nxpm/cli": "4.14.0",
     "@tailwindcss/custom-forms": "0.2.1",
     "@tailwindcss/forms": "0.2.1",
     "@types/jest": "27.4.1",
-    "@types/node": "14.14.33",
+    "@types/node": "18.0.0",
     "@typescript-eslint/eslint-plugin": "4.15.2",
     "@typescript-eslint/parser": "4.15.2",
     "cypress": "6.5.0",
@@ -85,16 +85,16 @@
     "eslint-plugin-cypress": "2.11.2",
     "husky": "5.1.1",
     "jest": "27.5.1",
-    "jest-preset-angular": "11.1.1",
+    "jest-preset-angular": "11.1.2",
     "lint-staged": "10.5.4",
     "postcss": "8.2.6",
     "postcss-ng-tailwind-in-components": "0.0.3",
     "prettier": "2.8.4",
     "tailwindcss": "2.0.3",
     "ts-jest": "27.1.4",
-    "ts-node": "9.1.1",
-    "typescript": "4.6.4",
-    "nx": "14.1.9"
+    "ts-node": "10.8.2",
+    "typescript": "4.7.4",
+    "nx": "14.4.3"
   },
   "lint-staged": {
     "nx.json": [
